refactor: improve the performance of Slider
