+++ src/Track.jsx
@@ -1,9 +1,10 @@
 import React from 'react';
 
-const Track = ({className, offset, length}) => {
+const Track = ({className, included, offset, length}) => {
   const style = {
     left: offset + '%',
     width: length + '%',
+    visibility: included ? 'visible' : 'hidden',
   };
   return <div className={className} style={style} />;
 };
